+++ BeGeneric.Backend.Abstractions/Exceptions/SecurityStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace BeGeneric.Backend.Common.Exceptions;
+
+public enum SecurityStatus
+{
+    Ok,
+    NotFound,
+    Unauthorised,
+    Forbidden,
+    BadRequest,
+    Conflict,
+}
