+++ BeGeneric.Backend.Abstractions/IComparerObject.cs
@@ -1,0 +1,8 @@
+﻿namespace BeGeneric.Backend.Common;
+
+public interface IComparerObject: IComparerObjectGroup
+{
+    string? Operator { get; set; }
+    string? Property { get; set; }
+    object? Filter { get; set; }
+}
