+++ BeGeneric.Backend.Abstractions/Models/ColumnMetadata.cs
@@ -1,0 +1,10 @@
+﻿namespace BeGeneric.Backend.Common.Models;
+
+public class ColumnMetadata
+{
+    public string TableName { get; set; }
+    public string ColumnName { get; set; }
+
+    public string AllowedValues { get; set; }
+    public string Regex { get; set; }
+}
