+++ BeGeneric.Backend.Abstractions/Models/RelatedEntityObject.cs
@@ -1,0 +1,6 @@
+﻿namespace BeGeneric.Backend.Common.Models;
+
+public class RelatedEntityObject<T>
+{
+    public T Id { get; set; }
+}
