+++ BeGeneric.Backend.Abstractions/Models/Role.cs
@@ -1,0 +1,6 @@
+﻿namespace BeGeneric.Backend.Common.Models;
+
+public class Role
+{
+    public string RoleName { get; set; }
+}
