+++ BeGeneric.Backend.Abstractions/Models/SummaryRequestObject.cs
@@ -1,0 +1,7 @@
+﻿namespace BeGeneric.Backend.Common.Models;
+
+public class SummaryRequestObject
+{
+    public string Property { get; set; }
+    public string AggregationType { get; set; }
+}
