+++ BeGeneric.Backend.Abstractions/Models/SummaryTypes.cs
@@ -1,0 +1,11 @@
+﻿namespace BeGeneric.Backend.Common.Models;
+
+public static class SummaryTypes
+{
+    public const string SUM = "sum";
+    public const string AVG = "avg";
+    public const string MAX = "max";
+    public const string MIN = "min";
+    public const string COUNT = "count";
+    public const string COUNT_DISTINCT = "count_distinct";
+}
