--- BeGeneric.Backend/GenericModels/EndpointProperty.cs
@@ -1,21 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace BeGeneric.Backend.Models
-{
-    public class EndpointProperty
-    {
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        [Key]
-        public Guid EndpointPropertyId { get; set; }
-
-        public Guid EndpointId { get; set; }
-
-        public string PropertyName { get; set; }
-        public string PropertyPath { get; set; }
-        
-
-        public virtual Endpoint Endpoint { get; set; }
-    }
-
-}
