--- BeGeneric.Backend/GenericModels/Role.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace BeGeneric.Backend.Models
-{
-    public class Role
-    {
-        public string RoleName { get; set; }
-    }
-}
