+++ BeGeneric.Backend/Services/GenericBackend/DataRequestObject.cs
@@ -1,6 +1,6 @@
-﻿namespace BeGeneric.Backend.Services.BeGeneric
+﻿namespace BeGeneric.Backend.Services.GenericBackend
 {
-    public class DataRequestObject: ComparerObject
+    public class DataRequestObject : ComparerObject
     {
         public SummaryRequestObject[]? Summaries { get; set; }
     }
