--- BeGeneric.Backend/Services/GenericBackend/DatabaseStructure/IDatabaseStructureService.cs
@@ -1,28 +1,0 @@
-﻿namespace BeGeneric.Backend.Services.BeGeneric.DatabaseStructure
-{
-    public interface IDatabaseStructureService
-    {
-        Type GetFieldType(string fieldName, string tableName);
-
-        string[] GetEnumValues(string fieldName, string tableName);
-
-        DatabaseFieldSizeLimitation GetFieldSizeLimitation(string fieldName, string tableName);
-        bool GetFieldNullable(string propertyName, string tableName);
-
-        string GetRegexValues(string fieldName, string tableName);
-
-        string ColumnDelimiterLeft { get; }
-
-        string ColumnDelimiterRight { get; }
-
-        string StringDelimiter { get; }
-
-        string DataSchema { get; }
-    }
-
-    public class DatabaseFieldSizeLimitation
-    {
-        public int? Min { get; set; }
-        public int? Max { get; set; }
-    }
-}
