--- BeGeneric.Backend/Services/GenericBackend/Exceptions/SecurityStatus.cs
@@ -1,12 +1,0 @@
-﻿namespace BeGeneric.Backend.Services.BeGeneric.Exceptions
-{
-    public enum SecurityStatus
-    {
-        Ok,
-        NotFound,
-        Unauthorised,
-        Forbidden,
-        BadRequest,
-        Conflict,
-    }
-}
