+++ BeGeneric.Backend/Services/GenericBackend/FieldValueErrorEnum.cs
@@ -1,4 +1,4 @@
-﻿namespace BeGeneric.Backend.Services.BeGeneric
+﻿namespace BeGeneric.Backend.Services.GenericBackend
 {
     public enum FieldValueErrorEnum
     {
