+++ BeGeneric.Backend/Services/GenericBackend/GenericDataService.cs
@@ -1,21 +1,18 @@
-﻿using BeGeneric.Backend.Models;
-using BeGeneric.Backend.Services.BeGeneric.DatabaseStructure;
-using BeGeneric.Backend.Services.BeGeneric.Exceptions;
+﻿using BeGeneric.Backend.Services.GenericBackend.Exceptions;
+using BeGeneric.Backend.Services.GenericBackend.Helpers;
 using BeGeneric.Backend.Settings;
-using BeGeneric.Helpers;
 using Microsoft.Extensions.Logging;
 using System.Data;
 using System.Data.Common;
 using System.Data.SqlClient;
-using System.Linq;
 using System.Security.Claims;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.RegularExpressions;
-using Endpoint = BeGeneric.Backend.Models.Endpoint;
+using Endpoint = BeGeneric.Backend.GenericModels.Endpoint;
 
-namespace BeGeneric.Backend.Services.BeGeneric
+namespace BeGeneric.Backend.Services.GenericBackend
 {
     public class GenericDataService<T> : IGenericDataService<T>
     {
@@ -50,7 +47,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             entities = entityDefinitions.ProcessEntities();
             this.logger = logger;
             this.dbStructure = dbStructure;
-            this.dbSchema = dbStructure.DataSchema;
+            dbSchema = dbStructure.DataSchema;
             this.connection = connection;
             this.attachedActionService = attachedActionService;
         }
@@ -62,7 +59,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             string userName = user.Identity.IsAuthenticated ? (user.Identity as ClaimsIdentity).FindFirst("id").Value : null;
             string roleName = user.Identity.IsAuthenticated ? (user.Identity as ClaimsIdentity).FindFirst(ClaimsIdentity.DefaultRoleClaimType).Value : null;
 
-            var action = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Get, ActionOrderType.Before);
+            var action = attachedActionService.GetAttachedAction(controllerName, ActionType.Get, ActionOrderType.Before);
             if (action != null)
             {
                 await action(new ActionData<T>()
@@ -109,7 +106,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 command.Parameters.AddRange(filters.Item3.Select(x => new SqlParameter(x.Item1, x.Item2)).ToArray());
             }
 
-            if (connection.State != System.Data.ConnectionState.Open)
+            if (connection.State != ConnectionState.Open)
             {
                 connection.Open();
             }
@@ -127,7 +124,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             if (success)
             {
                 string finalResponse = response.ToString();
-                var afterAction = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Get, ActionOrderType.After);
+                var afterAction = attachedActionService.GetAttachedAction(controllerName, ActionType.Get, ActionOrderType.After);
                 if (afterAction != null)
                 {
                     await afterAction(new ActionData<T>()
@@ -154,7 +151,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             string? userName = user.Identity.IsAuthenticated ? (user.Identity as ClaimsIdentity).FindFirst("id").Value : null;
             string? roleName = user.Identity.IsAuthenticated ? (user.Identity as ClaimsIdentity).FindFirst(ClaimsIdentity.DefaultRoleClaimType).Value : null;
 
-            var action = this.attachedActionService.GetAttachedAction(controllerName, ActionType.GetAll, ActionOrderType.Before);
+            var action = attachedActionService.GetAttachedAction(controllerName, ActionType.GetAll, ActionOrderType.Before);
             if (action != null)
             {
                 await action(new ActionData<T>()
@@ -253,7 +250,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                     command.Parameters.AddRange(filters.Item3.Select(x => new SqlParameter(x.Item1, x.Item2)).ToArray());
                 }
 
-                if (connection.State != System.Data.ConnectionState.Open)
+                if (connection.State != ConnectionState.Open)
                 {
                     connection.Open();
                 }
@@ -270,10 +267,10 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 ""recordsTotal"": {totalCount},
                 ""recordsFiltered"": {filteredTotalCount},
                 ""data"": {(entitiyList.Length == 0 ? "[]" : entitiyList)},
-                ""aggregation"": { (summaries != null ? ("[" + string.Join(", ", await AggregateEntityAccess(user, entity, null, filterObjectWithPermissions, summaries)) + "]") : "[]")}
+                ""aggregation"": {(summaries != null ? "[" + string.Join(", ", await AggregateEntityAccess(user, entity, null, filterObjectWithPermissions, summaries)) + "]" : "[]")}
             }}";
 
-            var afterAction = this.attachedActionService.GetAttachedAction(controllerName, ActionType.GetAll, ActionOrderType.After);
+            var afterAction = attachedActionService.GetAttachedAction(controllerName, ActionType.GetAll, ActionOrderType.After);
             if (afterAction != null)
             {
                 await afterAction(new ActionData<T>()
@@ -296,7 +293,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
         {
             (Entity entity, string permissionsFilter, string _) = await Authorize(user, endpoint.StartingEntity.ControllerName, getAll: true);
 
-            var action = this.attachedActionService.GetAttachedAction(endpoint.EndpointPath, ActionType.GetAll, ActionOrderType.Before);
+            var action = attachedActionService.GetAttachedAction(endpoint.EndpointPath, ActionType.GetAll, ActionOrderType.Before);
             if (action != null)
             {
                 await action(new ActionData<T>()
@@ -366,7 +363,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                     command.Parameters.AddRange(filters.Item3.Select(x => new SqlParameter(x.Item1, x.Item2)).ToArray());
                 }
 
-                if (connection.State != System.Data.ConnectionState.Open)
+                if (connection.State != ConnectionState.Open)
                 {
                     connection.Open();
                 }
@@ -382,7 +379,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             string finalResult = entitiyList.Length == 0 ? "[]" : entitiyList;
 
-            var afterAction = this.attachedActionService.GetAttachedAction(endpoint.EndpointPath, ActionType.GetAll, ActionOrderType.After);
+            var afterAction = attachedActionService.GetAttachedAction(endpoint.EndpointPath, ActionType.GetAll, ActionOrderType.After);
             if (afterAction != null)
             {
                 await afterAction(new ActionData<T>()
@@ -407,7 +404,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             (Entity entity, string permissionsFilter, string userId) = await Authorize(user, controllerName, post: true);
 
-            var action = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Post, ActionOrderType.Before);
+            var action = attachedActionService.GetAttachedAction(controllerName, ActionType.Post, ActionOrderType.Before);
             if (action != null)
             {
                 await action(new ActionData<T>()
@@ -425,7 +422,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             var usedProperties = properties
                 .Where(prop => (values.ContainsKey((prop.ModelPropertyName ?? prop.PropertyName).ToLowerInvariant()) || !string.IsNullOrEmpty(prop.DefaultValue))
                     && (!string.IsNullOrEmpty(prop.DefaultValue) || prop.ReferencingEntityId != null ||
-                        (values[(prop.ModelPropertyName ?? prop.PropertyName).ToLowerInvariant()] as JsonValue) != null));
+                        values[(prop.ModelPropertyName ?? prop.PropertyName).ToLowerInvariant()] as JsonValue != null));
 
             StringBuilder queryBuilder = new();
             queryBuilder.AppendLine(@$"DECLARE @generated_keys table(id {dbTypeParsers[typeof(T)]});");
@@ -515,7 +512,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             try
             {
-                if (connection.State != System.Data.ConnectionState.Open)
+                if (connection.State != ConnectionState.Open)
                 {
                     connection.Open();
                 }
@@ -552,7 +549,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
                     var savedEntity = await Get(user, controllerName, newId);
 
-                    var afterAction = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Post, ActionOrderType.After);
+                    var afterAction = attachedActionService.GetAttachedAction(controllerName, ActionType.Post, ActionOrderType.After);
                     if (afterAction != null)
                     {
                         await afterAction(new ActionData<T>()
@@ -635,7 +632,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             try
             {
                 using DbCommand command = new SqlCommand(queryBuilder.ToString(), connection);
-                if (connection.State != System.Data.ConnectionState.Open)
+                if (connection.State != ConnectionState.Open)
                 {
                     connection.Open();
                 }
@@ -654,12 +651,12 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             (Entity entity, string permissionsFilter, string _) = await Authorize(user, controllerName, put: true);
 
-            var action = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Patch, ActionOrderType.Before);
+            var action = attachedActionService.GetAttachedAction(controllerName, ActionType.Patch, ActionOrderType.Before);
             if (action != null)
             {
                 await action(new ActionData<T>()
                 {
-                    Id = id ?? (T)stringParsers[typeof(T)]((values[entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName.ToLowerInvariant() ?? "id"].ToString())),
+                    Id = id ?? (T)stringParsers[typeof(T)](values[entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName.ToLowerInvariant() ?? "id"].ToString()),
                     InputParameterData = JsonSerializer.Serialize(fieldValues),
                     UserName = user.Identity.IsAuthenticated ? (user.Identity as ClaimsIdentity).FindFirst("id").Value : null,
                     Role = user.Identity.IsAuthenticated ? (user.Identity as ClaimsIdentity).FindFirst(ClaimsIdentity.DefaultRoleClaimType).Value : null,
@@ -775,7 +772,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 throw new GenericBackendSecurityException(SecurityStatus.BadRequest, errors.Select(x => new { Error = x.Item1, Property = x.Item2 }));
             }
 
-            if (connection.State != System.Data.ConnectionState.Open)
+            if (connection.State != ConnectionState.Open)
             {
                 connection.Open();
             }
@@ -839,7 +836,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             await transaction.CommitAsync();
 
-            var afterAction = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Patch, ActionOrderType.After);
+            var afterAction = attachedActionService.GetAttachedAction(controllerName, ActionType.Patch, ActionOrderType.After);
             if (afterAction != null)
             {
                 await afterAction(new ActionData<T>()
@@ -859,7 +856,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
         {
             (Entity entity, string permissionsFilter, string _) = await Authorize(user, controllerName, delete: true);
 
-            var action = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Delete, ActionOrderType.Before);
+            var action = attachedActionService.GetAttachedAction(controllerName, ActionType.Delete, ActionOrderType.Before);
             if (action != null)
             {
                 await action(new ActionData<T>()
@@ -895,14 +892,14 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             string oneEntry = await Get(user, controllerName, id);
 
-            if (connection.State != System.Data.ConnectionState.Open)
+            if (connection.State != ConnectionState.Open)
             {
                 connection.Open();
             }
 
             await command.ExecuteNonQueryAsync();
 
-            var afterAction = this.attachedActionService.GetAttachedAction(controllerName, ActionType.Delete, ActionOrderType.After);
+            var afterAction = attachedActionService.GetAttachedAction(controllerName, ActionType.Delete, ActionOrderType.After);
             if (afterAction != null)
             {
                 await afterAction(new ActionData<T>()
@@ -977,7 +974,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             using DbCommand command = new SqlCommand(queryBuilder.ToString(), connection);
 
-            if (connection.State != System.Data.ConnectionState.Open)
+            if (connection.State != ConnectionState.Open)
             {
                 connection.Open();
             }
@@ -1000,7 +997,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 if (crossValue is JsonArray crossValueArray && crossValueArray != null)
                 {
                     var tmp1 = crossValueArray.ToArray().Select(x => (Guid)x[entity.Properties.First(x => x.IsKey).CamelCaseName() ?? "id"]);
-                    
+
                     if (clearRelations)
                     {
                         if (!string.IsNullOrEmpty(crossEntity.ValidToColumnName))
@@ -1071,7 +1068,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 Array.Empty<string>();
 
             StringBuilder sortBuilder = new();
-            
+
             if (sortTable.Length > 1)
             {
                 sortBuilder.Append($"JSON_VALUE({dbStructure.ColumnDelimiterLeft}{sortTable[0]}{dbStructure.ColumnDelimiterRight}, '$");
@@ -1172,20 +1169,20 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 command.Parameters.AddRange(filters.Item3.Select(x => new SqlParameter(x.Item1, x.Item2)).ToArray());
             }
 
-            if (connection.State != System.Data.ConnectionState.Open)
+            if (connection.State != ConnectionState.Open)
             {
                 connection.Open();
             }
 
             if (summaries == null)
             {
-                return new string[] { ((await command.ExecuteScalarAsync() as int?) ?? 0).ToString() };
+                return new string[] { (await command.ExecuteScalarAsync() as int? ?? 0).ToString() };
             }
             else
             {
                 List<string> result = new();
                 using DbDataReader dr = await command.ExecuteReaderAsync();
-                if (!(await dr.ReadAsync()))
+                if (!await dr.ReadAsync())
                 {
                     return null;
                 }
@@ -1324,7 +1321,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
             StringBuilder queryBuilder = new();
             int internalCounter = ++counter;
 
-            queryBuilder.AppendLine($"SELECT tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight} AS {dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName.CamelCaseName() ?? "id"}{dbStructure.ColumnDelimiterRight} ");
+            queryBuilder.AppendLine($"SELECT tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight} AS {dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName.CamelCaseName() ?? "id"}{dbStructure.ColumnDelimiterRight} ");
 
             foreach (Property property in entity.Properties.Where(x => !x.IsKey && !x.IsHidden))
             {
@@ -1377,8 +1374,8 @@ namespace BeGeneric.Backend.Services.BeGeneric
                     whereQueryAddition += $@" AND ({dbStructure.ColumnDelimiterLeft}{relation.ValidToColumnName}{dbStructure.ColumnDelimiterRight} IS NULL OR {dbStructure.ColumnDelimiterLeft}{relation.ValidToColumnName}{dbStructure.ColumnDelimiterRight} >= GETDATE()) ";
                 }
 
-                queryBuilder.Append($", ({GenerateSelectQuery(relation.Entity2, entities.Union(new Guid[] { relation.Entity2Id }), ref counter, roleName, userName, parameters)} AND {dbStructure.ColumnDelimiterLeft}{relation.Entity2.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight} IN (SELECT {dbStructure.ColumnDelimiterLeft}{relation.Entity2ReferencingColumnName}{dbStructure.ColumnDelimiterRight} FROM " +
-                    $"{dbSchema}.{dbStructure.ColumnDelimiterLeft}{relation.CrossTableName}{dbStructure.ColumnDelimiterRight} WHERE {dbStructure.ColumnDelimiterLeft}{relation.Entity1ReferencingColumnName}{dbStructure.ColumnDelimiterRight} = tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight} {whereQueryAddition}) " +
+                queryBuilder.Append($", ({GenerateSelectQuery(relation.Entity2, entities.Union(new Guid[] { relation.Entity2Id }), ref counter, roleName, userName, parameters)} AND {dbStructure.ColumnDelimiterLeft}{relation.Entity2.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight} IN (SELECT {dbStructure.ColumnDelimiterLeft}{relation.Entity2ReferencingColumnName}{dbStructure.ColumnDelimiterRight} FROM " +
+                    $"{dbSchema}.{dbStructure.ColumnDelimiterLeft}{relation.CrossTableName}{dbStructure.ColumnDelimiterRight} WHERE {dbStructure.ColumnDelimiterLeft}{relation.Entity1ReferencingColumnName}{dbStructure.ColumnDelimiterRight} = tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight} {whereQueryAddition}) " +
                     $"FOR JSON AUTO, INCLUDE_NULL_VALUES) AS {dbStructure.ColumnDelimiterLeft}{relation.Entity1PropertyName.CamelCaseName()}{dbStructure.ColumnDelimiterRight}");
             }
 
@@ -1416,9 +1413,9 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 }
 
                 queryBuilder.Append($", ({GenerateSelectQuery(relation.Entity1, entities.Union(new Guid[] { relation.Entity1Id }), ref counter, roleName, userName, parameters)} AND " +
-                    $"{dbStructure.ColumnDelimiterLeft}{relation.Entity1.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight} IN " +
+                    $"{dbStructure.ColumnDelimiterLeft}{relation.Entity1.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight} IN " +
                     $"(SELECT {dbStructure.ColumnDelimiterLeft}{relation.Entity1ReferencingColumnName}{dbStructure.ColumnDelimiterRight} FROM {dbSchema}.{dbStructure.ColumnDelimiterLeft}{relation.CrossTableName}{dbStructure.ColumnDelimiterRight} " +
-                    $"WHERE {dbStructure.ColumnDelimiterLeft}{relation.Entity2ReferencingColumnName}{dbStructure.ColumnDelimiterRight} = tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight} {whereQueryAddition}) FOR JSON AUTO, INCLUDE_NULL_VALUES) " +
+                    $"WHERE {dbStructure.ColumnDelimiterLeft}{relation.Entity2ReferencingColumnName}{dbStructure.ColumnDelimiterRight} = tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight} {whereQueryAddition}) FOR JSON AUTO, INCLUDE_NULL_VALUES) " +
                     $"AS {dbStructure.ColumnDelimiterLeft}{relation.Entity2PropertyName.CamelCaseName()}{dbStructure.ColumnDelimiterRight}");
             }
 
@@ -1426,11 +1423,11 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             if (filterProperty != null)
             {
-                queryBuilder.AppendLine($"WHERE tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight}={filterProperty}");
+                queryBuilder.AppendLine($"WHERE tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight}={filterProperty}");
             }
             else if (filterValue != null)
             {
-                queryBuilder.AppendLine($"WHERE tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID" }{dbStructure.ColumnDelimiterRight}=@tab{internalCounter}_val");
+                queryBuilder.AppendLine($"WHERE tab{internalCounter}.{dbStructure.ColumnDelimiterLeft}{entity.Properties.FirstOrDefault(x => x.IsKey)?.PropertyName ?? "ID"}{dbStructure.ColumnDelimiterRight}=@tab{internalCounter}_val");
             }
             else
             {
@@ -1536,7 +1533,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                         (r.Post || !post) &&
                         (r.Put || !put) &&
                         (r.Delete || !delete))
-                    .OrderBy(x => (getAll || getOne) ? x.ViewFilter : x.EditFilter)
+                    .OrderBy(x => getAll || getOne ? x.ViewFilter : x.EditFilter)
                     .FirstOrDefault();
 
                 entity = er?.Entity ?? entity;
@@ -1558,7 +1555,7 @@ namespace BeGeneric.Backend.Services.BeGeneric
                 }
             }
 
-            string permissionsFilter = (getAll || getOne) ? er?.ViewFilter : er?.EditFilter;
+            string permissionsFilter = getAll || getOne ? er?.ViewFilter : er?.EditFilter;
 
             if (!string.IsNullOrEmpty(permissionsFilter))
             {
@@ -1605,21 +1602,5 @@ namespace BeGeneric.Backend.Services.BeGeneric
 
             return permissionsFilter;
         }
-    }
-
-    public class RelatedEntityObject<T>
-    {
-        public T Id { get; set; }
-    }
-
-    public record SelectPropertyData
-    {
-        public string? JoinTableName { get; set; }
-        public string? JoinPropertyName { get; set; }
-        public string? OriginalTableName { get; set; }
-        public string? TableName { get; set; }
-        public string? IdPropertyName { get; set; }
-        public string? TableDTOName { get; set; }
-        public List<Tuple<string, string, string>>? Properties { get; set; }
     }
 }
